backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
